Added support for missing features
