+++ database/types/relationships.go
@@ -2,8 +2,8 @@ package types
 
 // RelationshipRow represents a database row containing the data of a relationship between two users
 type RelationshipRow struct {
-	Sender   string `db:"sender"`
-	Receiver string `db:"receiver"`
+	Sender   string `db:"sender_address"`
+	Receiver string `db:"receiver_address"`
 	Subspace string `db:"subspace"`
 }
 
@@ -17,8 +17,8 @@ func (row RelationshipRow) Equal(other RelationshipRow) bool {
 
 // BlockageRow represents a single database row containing the data of a user blockage
 type BlockageRow struct {
-	Blocker  string `db:"blocker"`
-	Blocked  string `db:"blocked_user"`
+	Blocker  string `db:"blocker_address"`
+	Blocked  string `db:"blocked_user_address"`
 	Reason   string `db:"reason"`
 	Subspace string `db:"subspace"`
 }
