+++ schema/01-profiles.sql
@@ -8,3 +8,9 @@ CREATE TABLE profile
     cover_pic     TEXT                        NOT NULL DEFAULT '',
     creation_time TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT NOW()
 );
+
+CREATE TABLE dtag_transfer_requests
+(
+    sender_address   TEXT NOT NULL REFERENCES profile (address),
+    receiver_address TEXT NOT NULL REFERENCES profile (address)
+);
